+++ cinema/Sphere.cpp
@@ -4,7 +4,7 @@
 #include "Intersection.h"
 #include <cmath>
 
-Sphere::Sphere(const Point &center,const float radius):
+Sphere::Sphere(const Point &center,const double radius):
 	m_center(center),m_radius(radius),m_squareRadius(radius*radius)
 {
 }
@@ -14,15 +14,15 @@ Sphere::~Sphere()
 {
 }
 
-bool Sphere::Intersect(const Ray & ray, float * t, Intersection *intersection)
+bool Sphere::Intersect(const Ray & ray, double * t, Intersection *intersection)
 {
 	Vector v = ray.o - m_center;
-	float vDotD = Vector::Dot(v, ray.d);
-	float squareD = ray.d.Square();
-	float delta = vDotD*vDotD - squareD*v.Square() + squareD*m_squareRadius;
+	double vDotD = Vector::Dot(v, ray.d);
+	double squareD = ray.d.Square();
+	double delta = vDotD*vDotD - squareD*v.Square() + squareD*m_squareRadius;
 	if (delta >= 0)
 	{
-		float tHit = (-vDotD - std::sqrtf(delta)) / squareD;
+		double tHit = (-vDotD - std::sqrt(delta)) / squareD;
 		if (tHit > 0.0f)
 		{
 			*t = tHit;
