+++ cinema/Transform.h
@@ -6,20 +6,20 @@ class Transform
 public:
 	Transform();
 	Transform(
-		float a00, float a01, float a02, float a03,
-		float a10, float a11, float a12, float a13,
-		float a20, float a21, float a22, float a23,
-		float a30, float a31, float a32, float a33);
+		double a00, double a01, double a02, double a03,
+		double a10, double a11, double a12, double a13,
+		double a20, double a21, double a22, double a23,
+		double a30, double a31, double a32, double a33);
 	~Transform();
 	Vector operator()(const Vector &v)const;
 	Point operator()(const Point &p)const;
 	Transform operator*(const Transform &t)const;
 	void SetMatrix(
-		float a00, float a01, float a02, float a03,
-		float a10, float a11, float a12, float a13,
-		float a20, float a21, float a22, float a23,
-		float a30, float a31, float a32, float a33);
+		double a00, double a01, double a02, double a03,
+		double a10, double a11, double a12, double a13,
+		double a20, double a21, double a22, double a23,
+		double a30, double a31, double a32, double a33);
 private:
-	float m_matrix[4][4];
+	double m_matrix[4][4];
 };
